requirements.txt
Flask==3.1.2
gunicorn==22.0.0
python-dotenv==1.2.1
torch==2.8.0
torchvision==0.23.0
timm==1.0.20
pillow== 11.3.0
numpy==1.26.4
scikit-learn== 1.7.1
pandas== 2.3.2
matplotlib==3.10.6
mlflow==2.14.0
dagshub==0.6.3
seaborn== 0.13.2
tqdm==4.67.1

# DVC and its extensions
dvc==3.63.0
dvc-s3==3.2.2